delete model/utils
