+++ .editorconfig
@@ -1,5 +1,5 @@
 # http://editorconfig.org
-
+# 支持多个不同的IDE，如果使用的IDE不同，IDE默认设定编码格式也不同，editorconfig跟项目走，只要IDE支持editconfig就能保证所有编码风格统一，比如tab为2个空格。
 root = true
 
 [*]
