--- npm-debug.log.2531311612
@@ -1,24 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/Users/zhushengjia/Desktop/drive/projects/nvm/v6.5.0/bin/node',
-1 verbose cli   '/Users/zhushengjia/Desktop/drive/projects/nvm/v6.5.0/bin/npm',
-1 verbose cli   'config',
-1 verbose cli   'get',
-1 verbose cli   'strict-ssl' ]
-2 info using npm@3.10.3
-3 info using node@v6.5.0
-4 verbose exit [ 0, true ]
-5 verbose stack Error: write EPIPE
-5 verbose stack     at exports._errnoException (util.js:1026:11)
-5 verbose stack     at WriteWrap.afterWrite (net.js:795:14)
-6 verbose cwd /Users/zhushengjia/Desktop/drive/projects/f2e/websoft/mbook
-7 error Darwin 15.4.0
-8 error argv "/Users/zhushengjia/Desktop/drive/projects/nvm/v6.5.0/bin/node" "/Users/zhushengjia/Desktop/drive/projects/nvm/v6.5.0/bin/npm" "config" "get" "strict-ssl"
-9 error node v6.5.0
-10 error npm  v3.10.3
-11 error code EPIPE
-12 error errno EPIPE
-13 error syscall write
-14 error write EPIPE
-15 error If you need help, you may report this error at:
-15 error     <https://github.com/npm/npm/issues>
-16 verbose exit [ 1, true ]
